Allow for developement and deployment of whole project from Qt Creator
workspace.pro will load the dependencies for the groovy projcet, if
the exist as checkouts next to groovy.

The groovy.pro modifications allows for local linking, instead of
using pkgconfig. This way we can develop the libs without
requiring a make install. Also added deployment rules for the QML
files. These only work for when deploying to the N900 for now.